==> code/02b_data_prep_dor_footprint_heights.Rmd <==
---
title: "Prepare building footprints and heights for Dortmund"
output: html_document
---

## Load packages

```{r, echo=FALSE}
library(tidyverse)
library(sf)
```

## Set root directory

Set path to your wd in masterfile or directly here. 

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = normalizePath(my_wd))
```

## Load Data

* Case study region
* Building footprints 2011 and 2021: The data is already roughly reduced to the extent of the study area to reduce computing time. 
* Footprints of 3D building model with information on building heights 2022 (2021 was not available for this study)

```{r}
dor_case <- st_read("./output/dor_iso30_VWG.gpkg")

dor_bldg_raw_2021 <- st_read("./input/dor_hu_2021_raw.gpkg")

dor_bldg_raw_2011 <- st_read("./input/dor_hu_2011.gpkg")

dor_lod2_2022 <- st_read("./input/dor_lod2_2022_raw.gpkg")
```

## Filter to case study region

```{r}
st_crs(dor_case) == st_crs(dor_bldg_raw_2021)

dor_bldg_2021_case <- st_filter(dor_bldg_raw_2021, dor_case, .predicate = st_intersects)

dor_bldg_2011_case <- st_filter(dor_bldg_raw_2011, dor_case, .predicate = st_intersects)
```

## Join data building height and building function 2021/22

Building height can be joined using the bldg_gmlid. We join bldg_measured_height and bldg_function. 
The building function is later used to distinguish residential/non-residential buildings.  

```{r}
bldg_heights_clean <- dor_lod2_2022 %>% 
  st_drop_geometry() %>% 
  select(bldg_gmlid, bldg_measured_height,  bldg_function)

dor_bldg_2021_height <- dor_bldg_2021_case %>%
  left_join(bldg_heights_clean, by = join_by("oi" == "bldg_gmlid"))
```

## Write results

```{r}
st_write(dor_bldg_2021_height, "./output/dor_bldg_2021_01_ha_raw.gpkg", delete_layer = TRUE)
st_write(dor_bldg_2011_case, "./output/dor_bldg_2011_01_raw.gpkg", delete_layer = TRUE)
```
