==> code/02c_data_prep_str_footprint_heights.Rmd <==
---
title: "Prepare building footprints and heights for Strasbourg"
output: html_document
---

## Load packages

```{r, echo=FALSE}
library(tidyverse)
library(sf)
```

## Set root directory

Set path to your wd in masterfile or directly here. 

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = normalizePath(my_wd))
```

## Load Data

* Case study region
* Building footprints 2011 and 2022: We use building footprints of 2022 instead of 2021 (in 2022 all footprints are derived from cadastral data while in 2021 there are some commune which still contain footprints derived from aerial images - see Bucher et al. 2025 for further information)
* Building height is already included in building footprints

Data can be downloaded from https://geoservices.ign.fr/bdtopo#telechargementshpdep2022 and https://geoservices.ign.fr/bdtopo#telechargement2011. 
It is organized by departements. For the case of Strasbourg, Departement 67 is needed.

```{r}
str_case <- st_read("./output/str_iso30_commune_dissolved.gpkg")

bati_a_2011 <- st_read("./input/BD_Topo/2011/BDTOPO_2-0_TOUSTHEMES_SHP_LAMB93_D067_2011-04-12/BDTOPO/1_DONNEES_LIVRAISON_2021-10-00019/BDT_2-0_SHP_LAMB93_D067-ED111/E_BATI/BATI_INDIFFERENCIE.SHP")
bati_b_2011 <- st_read("./input/BD_Topo/2011/BDTOPO_2-0_TOUSTHEMES_SHP_LAMB93_D067_2011-04-12/BDTOPO/1_DONNEES_LIVRAISON_2021-10-00019/BDT_2-0_SHP_LAMB93_D067-ED111/E_BATI/BATI_INDUSTRIEL.SHP")
bati_c_2011 <- st_read("./input/BD_Topo/2011/BDTOPO_2-0_TOUSTHEMES_SHP_LAMB93_D067_2011-04-12/BDTOPO/1_DONNEES_LIVRAISON_2021-10-00019/BDT_2-0_SHP_LAMB93_D067-ED111/E_BATI/BATI_REMARQUABLE.SHP")

str_bldg_raw_2022 <- st_read("./input/BD_Topo/2022/BDTOPO_3-0_TOUSTHEMES_SHP_LAMB93_D067_2022-03-15/BDTOPO_3-0_TOUSTHEMES_SHP_LAMB93_D067_2022-03-15/BDTOPO/1_DONNEES_LIVRAISON_2022-03-00081/BDT_3-0_SHP_LAMB93_D067-ED2022-03-15/BATI/BATIMENT.shp")
```

## Define further inputs

```{r}
reference_crs <- st_crs(bati_a_2011)
```

## Join building layers to one layer for each year

```{r}
str_bldg_raw_2011 <- bati_a_2011 %>% 
  bind_rows(bati_b_2011) %>% 
    bind_rows(bati_c_2011)
```

## Filter to case study region

```{r}
st_crs(str_case) == st_crs(str_bldg_raw_2022)
str_case <- st_transform(str_case, reference_crs)

str_bldg_2022_case <- st_filter(str_bldg_raw_2022, str_case, .predicate = st_intersects)

str_bldg_2011_case <- st_filter(str_bldg_raw_2011, str_case, .predicate = st_intersects)
```

## Write results

```{r}
st_write(str_bldg_2022_case, "./output/str_bldg_2022_01_ha_raw.gpkg", delete_layer = TRUE)
st_write(str_bldg_2011_case, "./output/str_bldg_2011_01_raw.gpkg", delete_layer = TRUE)
```
