==> code/06c_str_combine_bldg_metrics_and_training_data.Rmd <==
---
title: "Join trainingsdata random forest Strasbourg"
output: html_document
---

## Load packages

```{r, echo=FALSE}
library(tidyverse)
library(sf)
```

## Set root directory

Set path to your wd in masterfile or directly here. 

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = normalizePath(my_wd))
```

## Load Data

```{r, include=FALSE}
bldg_metrics <- st_read("./scripts/bldg_features/str_input_bldg_metrics.gdb", layer = "str_bldg_BLK_GA_BREG") %>% 
  st_drop_geometry()

str_bldg_22 <- st_read("./output/str_bldg_2022_02_filtered.gpkg")

data_train_raw <- read_sf("./input/bldg_classification/str_trainingdata_lucie_2505.gpkg") %>% 
  st_drop_geometry() 
```

## Explore data

```{r}
summary(bldg_metrics)
```

## Join data

```{r}
dat_train <- data_train_raw %>%
  mutate(REFSTR = description_v2, 
         REF = training_v2) %>% 
  st_drop_geometry() %>%
  group_by(REFSTR, REF) %>%
  summarise(n = n())

## filter to residential buildings only and simplify types
trainingsdata_prep <- data_train_raw %>% 
    mutate(REFSTR = description_v2, 
         REF = training_v2) %>% 
  mutate(REFSTR_simple = if_else(REF %in% c(31, 32, 33, 34, 35), "SFH", NA),
         REFSTR_simple = if_else(REF %in% c(12, 13, 21, 22, 23), "MFH", REFSTR_simple),
         REFSTR_simple = if_else(REF %in% c(99), "aux", REFSTR_simple),
         REFSTR_simple = if_else(REF %in% c(1, 41), "non_res", REFSTR_simple)) %>% 
  select(part_id, REF, REFSTR, REFSTR_simple)

data_fin <- str_bldg_22 %>% 
  st_drop_geometry() %>% 
  select(-Gebaeude_H, -area_foot, -peri_foot) %>% 
  left_join(trainingsdata_prep, by = join_by("ID" == "part_id")) %>% 
  left_join(bldg_metrics, by = join_by("ID" == "ID")) %>% 
  select(ID, area_foot, peri_foot, REFSTR, REFSTR_simple, Join_Count:BLK_Join_Count, Gebaeude_H:SCHUMM,
          NEAR_FID:MINDISTBLDG, block_id, OID_BLK, AREA_BLK, SHPXREG:BUILDUP)

data_fin %>%
  group_by(REFSTR_simple) %>%
  summarise(n = n())

colSums(is.na(data_fin))
```

## Write data

```{r}
write_rds(data_fin, "./output/str_bldg_input_rf.rds")
```
